# api-spec.yml
openapi: 3.0.2
info:
  title: 新歓スタンプラリーAPI
  version: '1.0'
  license:
    name: MIT
  contact:
    name: StudioAquatan
    url: 'https://www.aquatan.studio'
    email: f0reach@f0reach.me
servers:
  - url: 'http://localhost:3000'
paths:
  /user:
    post:
      summary: ユーザ作成
      operationId: post-user
      description: 匿名ユーザを作成し，認証用トークンを返します
      responses:
        '200':
          description: ユーザが正常に作成された
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: 認証トークン
                  user:
                    $ref: '#/components/schemas/User'
                required:
                  - token
                  - user
        '400':
          description: Bad Request
    get:
      summary: ユーザ情報
      operationId: get-user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
      parameters: []
      security:
        - Authorization: []
    patch:
      summary: ユーザ更新
      description: ユーザ情報を更新します．現在はニックネーム更新のみです
      operationId: patch-user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
      security:
        - Authorization: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                  maxLength: 15
                  nullable: true
                  description: 新しいニックネーム，nullでニックネーム削除
...
components:
  schemas:
    User:
      title: 利用者
      type: object
      description: スタンプラリー参加者の情報です
      properties:
        id:
          type: string
          description: ユーザーのUUID
          format: uuid
          readOnly: true
        nickname:
          type: string
          nullable: true
          description: OGP画像生成で使われるニックネームです．設定されていない場合nullになります
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: ユーザー登録の日時です
      required:
        - id
        - nickname
        - createdAt
...
  securitySchemes:
    Authorization:
      name: Authorization
      type: apiKey
      in: header